(tf) Remove libvirt pool/baseimage import from terraform manifests
The trouble with managing download and import of the base image and of
the separat volume pool directory via terraform is that the idempotency
of the resources is not handled well across invocations. The base image
and pool only need be set up once and then may be used any number of
times for various clusters of that platform, but terraform fails because
of the existence of the resources it is being asked to maintain and
instead would require an import step each time. Rather than try to
automate that route, I'm removing those resources and will manage them
separately as tasks using the ruby-libvirt gem for direct libvirt api
access. This will be cleaner than trying to scrape the human readable
tabular output of a cli tool like virsh.

* Remove base image url generation from platform string from
  terraform/locals.tf as it is no longer used within the terraform
  manifests and the functions/get_image_url.pp will provide for a
  plan to generate the same.
* Remove libvirt_pool and base libvirt_volume resources from terraform
  manifests. Instead, supply the base_volume_name and pool_name directly
  to the manifests as variables.
* Add a libvirt_network resource to isolate each test cluster within
  it's own ip range, provide dhcp and local dns for resolution of
  hostnames within the cluster.
* Change the cloud-init user-data name of the ssh login account to be
  configurable based on the var.user_name setting.