(plans) Add a subplan to manage download and import of base images
This replaces the terraform resources removed in an earlier commit
for the management of base image and platform pool. This subplan will
idempotently manage these for a given platform string assuming
the get_image_url() function is updated appropriately.

The spec_helper.rb is expanded to provide for bolt_spec support for plan
specs with a workaround for the modulepath muddle between rspec-puppet
and bolt_spec.