(bolt) Retrieve bolt inventory from a given terraform state file
Provides a template for the creation of a per cluster inventory.yaml
tied to the cluster's tfstate file. This will allow the
resolve_terraform_targets() function to extract functional targets by
role based on keywords in the hostname. This is a bit of a workaround
for the issue noted in the template regarding the use of terraform
modules and the current limits of the puppetlabs-terraform
resolve_references() function.