(plans) Add a subplan to install puppet on the cluster
...using the openvox_bootstrap module.

The Bolt apply_prep() function implicitly makes use of the
puppetlabs-puppet_agent modules puppet_agent::install task to install an
agent on a node from the Perforce apt/yum.puppet.com repositories. The
collections to be installed are locked by an Enum that does not include
the openvox7/8 collection names.

But control of which task to use for this function is based on Bolt's
puppet_library plugin hook. So that plugin-hook is updated in the
bolt-project.yaml to make use of the openvox_bootstrap::install task,
which has been separately written in an openvox_bootstrap module.

Right now, the new install_puppet subplan does nothing more than
apply_prep to get puppet installed on the cluster nodes.