(tf) Ensure that the user_password variable defaults to ''
If null is passed explicitly (which it will be if the plan user_password
parameter is not set) without nullable false, then it gets passed to the
terraform templatefile() function as a null and fails because the
function won't interpolate a null value. With nullable=false the default
of '' is set instead when a null user_password is explicitly passed to
the terraform modules.