(plans) Add a plan for nested virtualization testing
Implements and documents the steps needed to prepare an existing vm for
nested libvirt virtualization so that it in turn can run a plan like
standup_cluster to set up vms within itself.

Mostly was using this to troubleshoot similar problems when attempting
to setup libvirt on github action servers for plan testing.

The most notable difference is that you can't reboot a GHA node.