(tasks) Allow create pool task to generate pools with arbitrary paths
The task previously created all pools as a subdirectory of the base
storage pool based on the name of the pool. This is still a default
behavior, but the task and library now allow explicitly setting either a
relative or absolute path for the target.

This in turn allows the prep vm plan to generate the libvirt default
storage pool for the default /var/lib/libvirt/images path itself when
bootstrapping.