(net) Simplify vm domain to just 'vm'
...from "${cluster-id}.vm". The simple libnss-libvirt package I'm using
for local dns resolution of generated vms just maps vm name to ip. So in
the old scheme, two separate clusters with a 'primary' domain, clash,
and only the last defined gets resolution. Consequently, the previous
comment made the hostnames unique by combining
"${cluster_id}-${role}-${count}" as the hostname, but this left a
redundant "${cluster_id}-${role}-${count}.${cluster_id}.vm" as the fqdn.
This change gives us "${cluster_id}-${role}-${count}.vm" as the fqdn.

Also reorganizes the standup_cluster params to move required parameters
to the top of the list.