(plans) Loosen os parameter requirements for standup_cluster
Since the os params can be specified for each Vm_spec, it's no longer
required to set them at the top level, though that's generally going to
be convenient.

Validates two general cases:
* must set all top level os params if set any
* must set all top level os params if any vm_spec is missing os params

Does not try to to validate details of partially specified os params in
the vm_specs, yet.