(plans) Add a subplan to lookup target platform from facts
Used to set a platform variable on each target based on the target
facts. Targets without facter (before puppet is installed) return a
basic set of os facts which does not include architecture. The plan
turns to `uname -m` for that case.

This will be used to obtain platform strings for beaker host files for
running acceptance tests, for example.