(plans) Update the install_openvox subplan with version parameters
The apply_prep function does not allow any parameters for version or
collection. This commit updates install_openvox to make use of
parameters for openvox collection and version, and whether to install a
released version or a pre-release build. Calls out directly to
openvox_bootstrap::install or openvox_bootstrap::install_build_artifact
tasks.

Sets the puppet-agent feature on the targets and collects facts
afterwards.