(gha) Provide defaults for collection/version in action.yaml
The inputs have defaults, but that doesn't handle the case of a calling
workflow passing an empty string to either collection or version. This
would throw an error from the
Kvm_automation_tooling::Openvox_install_params type. Providing defaults
here to the env vars before constructing the standup_cluster_params.json
is a simple way to ensure the install_openvox_params structure is
well-formed.