(plans,gha) Simplify cluster_id
Previously, the standup_cluster plan took a string parameter, 'cluster_name',
and generated a cluster_id from this, the puppet architecture, and os info to
generate a cluster_id like:

foo-singular-ubuntu-2404-amd64

This made the uniqueness of a given cluster_id platform dependent, but
meant that downstream tooling then needed to interpete pecularities of
platform string generation (ubuntu version handling, debian arch
translation, etc.) when determing a cluster_id. It also made hostnames
cumbersome since the generated hostnames were of the form:

{cluster_id}-{vm_role}-{count}

And this had a further problem in that a platform string such as
rocky-9-x86_64 is not a valid part of a hostname given the underscore.

At any rate, this commit simplifies so that standup_cluster takes
a simple cluster_id string directly, and the caller is responsible for
its uniqueness.

So a standup_cluster plan with a cluster_id of 'foo' will have:

* hostnames:
** foo-{vm_role}-{count}
* inventory:
** terraform/intances/inventory.foo.yaml
* terraform state:
** terraform/instances/foo.tfstate
** terraform/instances/foo.tfvars.json