(plans) Retry terraform until we get valid ipv4 addresses
There is a race cropping up on Redhat (rocky/alma) vms where the
terraform dmacvicar/libvirt provider queries qemu-agent for network
address and gets back an ipv6 before an ipv4 is set. Since it's a
legitmate address, it considers its job done and moves on, despite the
libvirt_domain.network_interface.wait_for_lease setting, but a
Bolt::Target.uri seems to need to be an ipv4, so generating the targets
later when we resolve_references fails.

To work around this, since terraform::apply is idempotent, I've wrapped
that in a ctrl::do_until call so that we retry a few times to get all
ipv4.

A failing terraform::apply plan should still raise an error and break
out, I believe.