(tasks) Fail download_image task for 400+ http responses
Curl without --fail or --fail-with-body stores the http response for 400
response like 404 not found in the -o file and exits 0. This would lead
to a silent failure where the pipeline would later grind to a halt after
libvirt attempts to stand up an image consisting of the http doc for a
404 response, and then terraform complains that their is no
networking...

Anyway, adding --fail-with-body captures the response and exits
non-zero, and the task reads that, clears the file and raises an error
with the curl output and the http response.

Sadly this doesn't necessarily print nicely wrapped in Bolt's parallelize(),
but it does fail, and the output does read nicely with --verbose. But
with --stream the PlanResult gets dumped as unformatted json.