(plans) Graduate install_openvox to a main plan
The standup_cluster plan is useful for any general vm cluster
specification and doesn't need to be tied to installing openvox
artifacts.

Likewise, installing openvox artifacts is going to become more
complicated with the addition of handling openvox-server, openvoxdb and
postgresql and it doesn't make sense to burden the standup_cluster plan's
parameters with a subhash of install details meant for the
install_openvox plan.

I'm leaving the basic install hook in standup_cluster for now, but have
moved the install_openvox plan from plans/subplans to plans/ and will
revise it so that it's straightforward to call manually as a second
automation step with a generated inventory from a previous standup_cluster
invocation.