(plans) Install openvox server packages
Revises the install_openvox plan to handle installation of
openvox-server and openvoxdb packages as well as the agent.

Updates standup_cluster plan to constrain the install_openvox_params to
match. Passing a complex hash of params for the subbordinate plan is a
convenience for running just standup_cluster from a params json file.

The local action.yaml is updated as well to set just the agent params.
I'm not intending on complicating the action with server/db version
params, and may go the other route and remove openvox installation from
the action altogether, allowing it to just be for vm setup.