(maint) Some puppet-lint fixes
Adding the puppet-lint gem and the vox-pupuli plugin dependencies. The
later includes the strict_indent plugin, which I'm including so I can
turn it off for now in the .puppet-lint.rc. The tool throws an error if
you try to include a switch for a feature that doesn't exist, which
would be fine if the LSP wasn't then trying to use puppet-lint alongside
whatever plugins have landed in the general gem pool from other project
dependencies.

Adds a puppet-lint to the GHA pr workflow as well.