(plans) Install openvoxdb-termini on server targets
The openvoxdb-termini package is built from the same openvoxdb project
source as the openvoxdb package, but contains Puppet functions, faces
and terminus classes for interacting with openvoxdb that are used by the
openvox-server (or typically on the cli of the server node).

This patch modifies the kvm_automation_tooling::install_openvox plan
to ensure that the openvoxdb-termini matching the $openvox_db_params
version (or $install_defaults version) will be installed on the
$openvox_server_targets alongside openvox-server.

An $install_termini flag can be set false to skip this behavior.