(plans) Install rbenv in prep_vm_for_module_testing
Installs rbenv instead of system ruby when preparing a vm to be able to
run kvm_automation_tooling within the vm. Slower, but ensures I can pick
a ruby version to test with. This makes for a better environment match
when using this plan to troubleshoot problems seen with nested_vms run
on gha ubuntu 22.04 hosts, for example.