(plans) Install prereqs for openvox-server and openvoxdb
Pre-release packages are downloaded from an artifacts server and
installed directly with rpm or dpkg. Consequently, package dependencies
are not automatically installed.

The install_openvox plan is updated here to install those dependencies
when it detects we're installing a pre-release version of openvox-server
or openvoxdb which depend on a jdk installation, among other things.

Some general refactoring of the plan makes this a little smoother to
read.

* Installing a single openvox package is extracted out as
  subplans/install_component.pp.
* Installing the pre-requisites is added as
  subplans/install_server_prerequisites.pp
* Transforming and aggregating the package version information by host
  is extracted out as transform_openvox_host_version_results() function