(plans) Add database to roles given to beaker host primary
The beaker hosts file has an archaic set of roles that translate via
cursed beaker helpers into magical collections that can be referenced by
acceptance suites wanting to perform actions on hosts that match a
certain role.

The 'master' role roughly corresponds to node that has a puppetserver on
it. The 'database' role roughly corresponds to node that has a puppetdb
on it. Now that I'm using the generate_beaker_hosts_file plan for
openvoxdb testing, the database role needs to be added because that
acceptance suite expects to work with database hosts rather than just
master.

At the moment, I'm just testing with simple layouts that have a single
primary node with all the services (openvox-server, openvoxdb,
postgresql...)