(plans) Drop unused allowed_platforms type
The Allowed_platforms type had been gateing platforms in the
get_image_url() function prior to 6afd292 when I switched to Os_spec.
After that Allowed_platforms was no longer used, but I forgot to remove
it at the time.

With Os_spec, platforms are gated by the Operating_system type which is
not version specific.