(plans,tf) Support arm64 vms
The standup_cluster plan now looks up host architecture and passes that
in to the terraform modules.

Terraform is updated to automatically determine domain.type based on
whether host and vm architecture match, allowing kvm, otherwise qemu is
chosen as type.

The modules now set the os arch and machine according the vm
architecture, along with features and cpu mode and model.

A domain_type parameter has been added to vm_spec to allow domain.type
to be overridden in case kvm is not available on the runner, for
example.

os.firmware is set to uefi now if the vm is arm64. Ubuntu images needed
this.